// app/src/generated-api/models/Driver.ts
export interface Driver {
    /**
     * ID único gerado pelo sistema.
     * @type {string}
     * @memberof Driver
     */
    id: string;
    /**
     * Nome do motorista.
     * @type {string}
     * @memberof Driver
     */
    name: string;
    /**
     * Número do CPF (apenas dígitos).
     * @type {string}
     * @memberof Driver
     */
    cpf?: string;
    /**
     * Número do RG (ex: MG 13.569.569).
     * @type {string}
     * @memberof Driver
     */
    rg?: string;
    /**
     * Data de Nascimento (formato YYYY-MM-DD).
     * @type {Date}
     * @memberof Driver
     */
    dateBirth: Date;
    /**
     * Número da Carteira Nacional de Habilitação (CNH).
     * @type {string}
     * @memberof Driver
     */
    cnhNumber: string;
    /**
     * Data de validade da CNH.
     * @type {Date}
     * @memberof Driver
     */
    validityCnh: Date;
    /**
     * Data de validade do exame toxicológico.
     * @type {Date}
     * @memberof Driver
     */
    validityToxicological: Date;
    /**
     * Timestamp de criação do registro.
     * @type {Date}
     * @memberof Driver
     */
    createdAt?: Date;
}